+++ database/migrations/1685249009398-AddProfilePictureURLToAdmin.ts
@@ -4,14 +4,30 @@ export class AddProfilePictureURLToAdmin1685249009398
   implements MigrationInterface
 {
   public async up(queryRunner: QueryRunner): Promise<void> {
-    // add a "profile_picture_url" column to the admin table
+    const hasProfilePictureURLColumn = await queryRunner.hasColumn(
+      "admin",
+      "profile_picture_path"
+    );
+
+    if (hasProfilePictureURLColumn) {
+      return;
+    }
+
     await queryRunner.query(
       `ALTER TABLE "admin" ADD COLUMN "profile_picture_path" TEXT NOT NULL DEFAULT ''`
     );
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
-    // drop the "profile_picture_url" column from the admin table
+    const hasProfilePictureURLColumn = await queryRunner.hasColumn(
+      "admin",
+      "profile_picture_path"
+    );
+
+    if (!hasProfilePictureURLColumn) {
+      return;
+    }
+
     await queryRunner.query(
       `ALTER TABLE "admin" DROP COLUMN "profile_picture_path"`
     );
